1.1 push - fixes for uptime 7.8.2 - updated libcurl
libcurl 7.50+ changes the way certificate information gets presented. it
used to be a subarray each for subject and issuer, now its LDAP style in
a single element of the array each. Lots of regex string replacement and
splitting was required. It's not perfect, but you get monitorable
results in subject and issuer detail. The plugin also works on older
uptime monitoring stations with older version of CURL. Also split up
content into separate content and encoding fields.